refactor(cas): move the label rules to shared/, out of accounts
webauthn's PasskeyName was reaching into accounts for the display name
sanitising and validation. The rules are about a user-facing label, not
about accounts, so they now live in src/shared/label.rs (LabelError,
MAX_LABEL_LENGTH, sanitize_label, validate_label) with their tests.
DisplayName and PasskeyName stay in their contexts as thin newtypes with
their own error types, so invalid_display_name and invalid_passkey_name
stay next to the handlers that map them.

LAYOUT.md gains the shared vocabulary layer and its bar for entry: at
least two contexts, none of them the owner, no axum, no SQL.
