+++ apps/cas/src/lib.rs
@@ -6,6 +6,7 @@ pub mod db;
 pub mod http;
 pub mod migrations;
 pub mod sessions;
+pub mod shared;
 #[cfg(test)]
 pub(crate) mod testing;
 pub mod webauthn;
